Move code from ecosystem package to explore package
