+++ .changeset/two-foxes-exercise.md
@@ -1,0 +1,6 @@
+---
+'@backstage/plugin-explore': patch
+---
+
+Rework the explore plugin to allow the user to explore things in the ecosystem,
+including tools and domains.
