--- plugins/ecosystem/.eslintrc.js
@@ -1,3 +1,0 @@
-module.exports = {
-  extends: [require.resolve('@backstage/cli/config/eslint')],
-};
