--- plugins/ecosystem/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@backstage/plugin-ecosystem",
-  "version": "0.1.1",
-  "main": "src/index.ts",
-  "types": "src/index.ts",
-  "license": "Apache-2.0",
-  "publishConfig": {
-    "access": "public",
-    "main": "dist/index.esm.js",
-    "types": "dist/index.d.ts"
-  },
-  "scripts": {
-    "build": "backstage-cli plugin:build",
-    "start": "backstage-cli plugin:serve",
-    "lint": "backstage-cli lint",
-    "test": "backstage-cli test",
-    "diff": "backstage-cli plugin:diff",
-    "prepack": "backstage-cli prepack",
-    "postpack": "backstage-cli postpack",
-    "clean": "backstage-cli clean"
-  },
-  "dependencies": {
-    "@backstage/catalog-model": "^0.6.1",
-    "@backstage/core": "^0.4.4",
-    "@backstage/plugin-catalog": "^0.2.10",
-    "@backstage/theme": "^0.2.2",
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "4.0.0-alpha.45",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "6.0.0-beta.0",
-    "react-use": "^15.3.3"
-  },
-  "devDependencies": {
-    "@backstage/cli": "^0.4.6",
-    "@backstage/dev-utils": "^0.1.7",
-    "@backstage/test-utils": "^0.1.6",
-    "@testing-library/jest-dom": "^5.10.1",
-    "@testing-library/react": "^10.4.1",
-    "@testing-library/user-event": "^12.0.7",
-    "@types/jest": "^26.0.7",
-    "@types/node": "^12.0.0",
-    "msw": "^0.21.2",
-    "cross-fetch": "^3.0.6"
-  },
-  "files": [
-    "dist"
-  ]
-}
