+++ plugins/explore/README.md
@@ -1,7 +1,22 @@
-# Title
+# explore
 
 Welcome to the explore plugin!
+This plugin helps to visualize the domains and in your ecosystem.
 
-## Sub-section 1
+## Getting started
 
-## Sub-section 2
+To install the plugin, include the following import your `plugins.ts`:
+
+```typescript
+export { plugin as Ecosystem } from '@backstage/plugin-explore';
+```
+
+Add a link to the sidebar in `Root.tsx`:
+
+```typescript
+import LayersIcon from '@material-ui/icons/Layers';
+
+...
+
+<SidebarItem icon={LayersIcon} to="explore" text="Explore" />
+```
