+++ plugins/explore/package.json
@@ -31,6 +31,7 @@
   },
   "dependencies": {
     "@backstage/core": "^0.5.0",
+    "@backstage/catalog-model": "^0.7.0",
     "@backstage/theme": "^0.2.2",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
@@ -38,7 +39,7 @@
     "classnames": "^2.2.6",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-router": "6.0.0-beta.0",
+    "react-router-dom": "6.0.0-beta.0",
     "react-use": "^15.3.3"
   },
   "devDependencies": {
