+++ plugins/explore/src/components/ExplorePage/index.ts
@@ -13,10 +13,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { plugin } from './plugin';
-
-describe('ecosystem', () => {
-  it('should export plugin', () => {
-    expect(plugin).toBeDefined();
-  });
-});
+export { ExplorePage } from './ExplorePage';
