+++ plugins/explore/src/routes.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 Spotify AB
+ * Copyright 2020 Spotify AB
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,5 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import '@testing-library/jest-dom';
-import 'cross-fetch/polyfill';
+
+import { createRouteRef } from '@backstage/core';
+
+const NoIcon = () => null;
+
+export const exploreRouteRef = createRouteRef({
+  icon: NoIcon,
+  path: '',
+  title: 'Explore',
+});
